ysi-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.26.0
redis==5.0.1
celery==5.3.6
pytest==7.4.4
pytest-asyncio==0.23.3
python-dotenv==1.0.0
email-validator==2.1.0
jinja2==3.1.3
python-dateutil==2.8.2
openai==1.9.0
langchain==0.1.1
pandas==2.1.4
numpy==1.26.3

# Vector database and embeddings
pgvector==0.2.4
sentence-transformers==2.2.2
tiktoken==0.5.2

# Full-text search
sqlalchemy-searchable==1.5.4

# Text processing
nltk==3.8.1
spacy==3.7.2

# Additional utilities for RAG
scikit-learn==1.4.0
faiss-cpu==1.7.4